Adding School District Boundaries
It turns out they have boundaries for the school districts in Allegheny
County as well! Also it should go without saying but make these
different colors from the municipality maps AND don't use both at the
same time. AND as it turns out, it just takes a minute for github to
update the files that I edit, so the two counties I changed the colors
of changed a little later in the day.