FIXED (SCHOOL SCHEMA DATA)
It was schema data! There was no name for the placemarks, so Google Maps
refused to take anything at all. So glad to have figured that out! Also
I'm making the municipality fills a transparent red so hopefully I can
clearly see both the municipality lines and the school districts as
well.